Fixed Javascript sidebar functions
Moved teh js file to the assets folder then called the file within the angular.json scripts section. Within the main.ts file the methods were declared and allowed access by any. These functions are called in a component withi the file that exports the functions to be used by the app.