Updated create story and chapter pages
updated teh create story and create chapter pages so that the tags appear on the create story page. Create chapter has the title appear as well and the data from both is displayed in the console. Create story also has the cover image input added in. Code taken fromt eh registration page image uploader. Tags are not adding to anything when clicked, but do changed color when pressed.